+++ workflow/scripts/build_demand.py
@@ -1,11 +1,7 @@
 """Builds the demand data for the PyPSA network."""
 
-# snakemake is not liking this futures import. Removing type hints in context class
-# from __future__ import annotations
-
 import calendar
 import logging
-import sqlite3
 import sys
 from abc import ABC, abstractmethod
 from pathlib import Path
@@ -20,7 +16,15 @@ import pypsa
 import xarray as xr
 from _helpers import configure_logging, get_multiindex_snapshots
 from constants_sector import FIPS_2_STATE, NAICS, VMT_UNIT_CONVERSION
-from eia import EnergyDemand, TransportationDemand
+from demand_scalers import (
+    AeoElectricityScaler,
+    AeoEnergyScalerApi,
+    AeoScaler,
+    AeoVmtScalerApi,
+    DemandScaler,
+    EfsElectricityScalar,
+)
+from eia import TransportationDemand
 
 logger = logging.getLogger(__name__)
 
@@ -1832,7 +1836,7 @@ class WriteIndustrial(WriteStrategy):
 
 
 ###
-# Formatters and Scalers
+# Formatters
 ###
 
 
@@ -1909,319 +1913,22 @@ class DemandFormatter:
     def assign_scaler(self):  # type DemandScaler
         """Assign logic to scale demand with."""
         if self.scaling_method == "aeo_energy":
-            assert self.api, "Must provide eia api key"
-            return AeoEnergyScaler(self.api)
+            assert self.filepath.endswith(".csv"), "Must provide formatted AEO.csv data"
+            return AeoScaler(self.filepath)
         elif self.scaling_method == "aeo_electricity":
             assert self.filepath.endswith(".sqlite"), "Must provide pudl.sqlite file"
             return AeoElectricityScaler(self.filepath)
         elif self.scaling_method == "efs":
             assert self.filepath.endswith(".csv"), "Must provide EFS.csv data"
             return EfsElectricityScalar(self.filepath)
-        elif self.scaling_method == "aeo_vmt":
+        elif self.scaling_method == "aeo_energy_api":
             assert self.api, "Must provide eia api key"
-            return AeoVmtScaler(self.api)
+            return AeoEnergyScalerApi(self.api)
+        elif self.scaling_method == "aeo_vmt_api":
+            assert self.api, "Must provide eia api key"
+            return AeoVmtScalerApi(self.api)
         else:
             raise NotImplementedError
-
-
-class DemandScaler(ABC):
-    """Allow the scaling of input data bases on different energy projections."""
-
-    def __init__(self):
-        self.projection = self.get_projections()
-
-    @abstractmethod
-    def get_projections(self) -> pd.DataFrame:
-        """Get implementation specific energy projections."""
-        pass
-
-    def get_growth(self, start_year: int, end_year: int, sector: str) -> float:
-        """Returns decimal change between two years."""
-        min_year = self.projection.index.min()
-        max_year = self.projection.index.max()
-
-        if start_year < min_year:
-            logger.warning(f"Setting base demand scaling year to {min_year}")
-            start_year = min_year
-        if end_year > max_year:
-            logger.warning(f"Setting final demand scaling year to {max_year}")
-            end_year = max_year
-
-        start = self.projection.at[start_year, sector]
-        end = self.projection.at[end_year, sector]
-
-        return end / start
-
-    def scale(
-        self,
-        df: pd.DataFrame,
-        start_year: int,
-        end_year: int,
-        sector: str,
-    ) -> pd.DataFrame:
-        """Scales data."""
-        growth = self.get_growth(start_year, end_year, sector)
-        new = df.mul(growth)
-        return self.reindex(new, end_year)
-
-    @staticmethod
-    def reindex(df: pd.DataFrame, year: int) -> pd.DataFrame:
-        """
-        Reindex a dataframe for a different planning horizon.
-
-        Input dataframe will be...
-
-        |                     | BusName_1 | BusName_2 | ... | BusName_n |
-        |---------------------|-----------|-----------|-----|-----------|
-        | 2019-01-01 00:00:00 |    aaa    |    ddd    |     |    ggg    |
-        | 2019-01-01 01:00:00 |    bbb    |    eee    |     |    hhh    |
-        | ...                 |    ...    |    ...    |     |    ...    |
-        | 2019-02-28 23:00:00 |    ccc    |    fff    |     |    iii    |
-
-        Output dataframe will be...
-
-        |                     | BusName_1 | BusName_2 | ... | BusName_n |
-        |---------------------|-----------|-----------|-----|-----------|
-        | 2030-01-01 00:00:00 |    aaa    |    ddd    |     |    ggg    |
-        | 2030-01-01 01:00:00 |    bbb    |    eee    |     |    hhh    |
-        | ...                 |    ...    |    ...    |     |    ...    |
-        | 2030-02-28 23:00:00 |    ccc    |    fff    |     |    iii    |
-        """
-        new = df.copy()
-        new.index = new.index.map(lambda x: x.replace(year=year))
-        return new
-
-
-class AeoElectricityScaler(DemandScaler):
-    """Scales against EIA Annual Energy Outlook electricity projections."""
-
-    def __init__(self, pudl: str, scenario: str = "reference"):
-        self.pudl = pudl
-        self.scenario = scenario
-        self.region = "united_states"
-        super().__init__()
-
-    def get_projections(self) -> pd.DataFrame:
-        """
-        Get sector yearly END-USE ELECTRICITY growth rates from AEO.
-
-        |      | power | units |
-        |----- |-------|-------|
-        | 2021 |  ###  |  ###  |
-        | 2022 |  ###  |  ###  |
-        | 2023 |  ###  |  ###  |
-        | ...  |       |       |
-        | 2049 |  ###  |  ###  |
-        | 2050 |  ###  |  ###  |
-        """
-        con = sqlite3.connect(self.pudl)
-        df = pd.read_sql_query(
-            f"""
-        SELECT
-        projection_year,
-        technology_description_eiaaeo,
-        gross_generation_mwh
-        FROM
-        core_eiaaeo__yearly_projected_generation_in_electric_sector_by_technology
-        WHERE
-        electricity_market_module_region_eiaaeo = "{self.region}" AND
-        model_case_eiaaeo = "{self.scenario}"
-        """,
-            con,
-        )
-
-        df = (
-            df.drop(columns=["technology_description_eiaaeo"])
-            .rename(
-                columns={"projection_year": "year", "gross_generation_mwh": "power"},
-            )
-            .groupby("year")
-            .sum()
-        )
-        df["units"] = "mwh"
-        return df
-
-
-class AeoEnergyScaler(DemandScaler):
-    """Scales against EIA Annual Energy Outlook energy projections."""
-
-    def __init__(self, api: str, scenario: str = "reference"):
-        self.api = api
-        self.scenario = scenario
-        self.region = "united_states"
-        super().__init__()
-
-    def get_sector_data(self, years: list[int], sector: str) -> pd.DataFrame:
-        """Function to piece togehter historical and projected values."""
-        start_year = min(years)
-        end_year = max(years)
-
-        data = []
-
-        if start_year < 2024:
-            data.append(
-                EnergyDemand(sector=sector, year=start_year, api=self.api).get_data(),
-            )
-        if end_year >= 2024:
-            data.append(
-                EnergyDemand(sector=sector, year=end_year, api=self.api).get_data(),
-            )
-        return pd.concat(data)
-
-    def get_projections(self) -> pd.DataFrame:
-        """
-        Get sector yearly END-USE ENERGY growth rates from AEO at a NATIONAL
-        level.
-
-        |      | residential | commercial  | industrial  | transport  | units |
-        |----- |-------------|-------------|-------------|------------|-------|
-        | 2018 |     ###     |     ###     |     ###     |     ###    |  ###  |
-        | 2019 |     ###     |     ###     |     ###     |     ###    |  ###  |
-        | 2020 |     ###     |     ###     |     ###     |     ###    |  ###  |
-        | ...  |             |             |             |            |       |
-        | 2049 |     ###     |     ###     |     ###     |     ###    |  ###  |
-        | 2050 |     ###     |     ###     |     ###     |     ###    |  ###  |
-        """
-        years = range(2017, 2051)
-
-        # sectors = ("residential", "commercial", "industry", "transport")
-        sectors = ("residential", "commercial", "industry")
-
-        df = pd.DataFrame(
-            index=years,
-        )
-
-        for sector in sectors:
-            sector_data = self.get_sector_data(years, sector).sort_index()
-            df[sector] = sector_data.value
-
-        df["units"] = "quads"
-        return df
-
-
-class AeoVmtScaler(DemandScaler):
-    """Scales against EIA Annual Energy Outlook vehicle mile traveled projections."""
-
-    def __init__(self, api: str, scenario: str = "reference"):
-        self.api = api
-        self.scenario = scenario
-        self.region = "united_states"
-        super().__init__()
-
-    def get_historical_value(self, year: int, sector: str) -> float:
-        """Returns single year value at a time."""
-        return TransportationDemand(vehicle=sector, year=year, api=self.api).get_data(pivot=True).values[0][0]
-
-    def get_future_values(
-        self,
-        year: int,
-        sector: str,
-    ) -> pd.DataFrame:
-        """Returns all values from 2024 onwards."""
-        return TransportationDemand(
-            vehicle=sector,
-            year=year,
-            api=self.api,
-            scenario=self.scenario,
-        ).get_data()
-
-    def get_projections(self) -> pd.DataFrame:
-        """
-        Get sector yearly END-USE ENERGY growth rates from AEO at a NATIONAL
-        level.
-
-        |      | light_duty | med_duty  | heavy_duty  | bus  | units |
-        |----- |------------|-----------|-------------|------|-------|
-        | 2018 |     ###    |    ###    |     ###     | ###  |  ###  |
-        | 2019 |     ###    |    ###    |     ###     | ###  |  ###  |
-        | 2020 |     ###    |    ###    |     ###     | ###  |  ###  |
-        | ...  |            |           |             |      |       |
-        | 2049 |     ###    |    ###    |     ###     | ###  |  ###  |
-        | 2050 |     ###    |    ###    |     ###     | ###  |  ###  |
-        """
-        years = range(2017, 2051)
-
-        vehicles = ("light_duty", "med_duty", "heavy_duty", "bus")
-
-        df = pd.DataFrame(
-            columns=["light_duty", "med_duty", "heavy_duty", "bus"],
-            index=years,
-        )
-
-        for year in sorted(years):
-            if year < 2024:
-                for vehicle in vehicles:
-                    df.at[year, vehicle] = self.get_historical_value(
-                        year,
-                        vehicle,
-                    )
-
-        for vehicle in vehicles:
-            aeo = self.get_future_values(max(years), vehicle)
-            for year in years:
-                if year < 2024:
-                    continue
-                df.at[year, vehicle] = aeo.at[year, "value"]
-
-        df["units"] = "thousand VMT"
-        return df
-
-
-class EfsElectricityScalar(DemandScaler):
-    """Scales against NREL Electrification Futures Study electricity projections."""
-
-    def __init__(self, filepath: str):
-        self.efs = filepath
-        self.region = "united_states"
-        super().__init__()
-
-    def read(self) -> pd.DataFrame:
-        """Read in raw EFS data."""
-        df = pd.read_csv(self.efs, engine="pyarrow")
-        return (
-            df.drop(
-                columns=[
-                    "Electrification",
-                    "TechnologyAdvancement",
-                    "LocalHourID",
-                    "Sector",
-                    "Subsector",
-                ],
-            )
-            .groupby(["Year", "State"])
-            .sum()
-        )
-
-    def interpolate(self, df: pd.DataFrame) -> pd.DataFrame:
-        """Function interpolates between provided demand data years."""
-        efs_years = df.index
-        new_years = range(min(efs_years), max(efs_years) + 1)
-        df = df.reindex(new_years)
-        return df.interpolate()
-
-    def get_projections(self) -> pd.DataFrame:
-        """
-        Get sector yearly END-USE ELECTRICITY growth rates from EFS. Linear
-        interpolates missing values.
-
-        |      | power | units |
-        |----- |-------|-------|
-        | 2018 |  ###  |  ###  |
-        | 2019 |  ###  |  ###  |
-        | 2020 |  ###  |  ###  |
-        | ...  |       |       |
-        | 2049 |  ###  |  ###  |
-        | 2050 |  ###  |  ###  |
-        """
-        df = self.read().reset_index()
-        if self.region == "united_states":
-            df = df.drop(columns="State").groupby("Year").sum()
-        else:
-            raise NotImplementedError
-        df = self.interpolate(df)
-        df = df.rename(columns={"LoadMW": "power"})
-        df["units"] = "MWh"
-        return df
 
 
 ###
@@ -2235,42 +1942,41 @@ def get_demand_params(
     **kwargs,
 ) -> tuple:
     """Gets hard coded demand options."""
-    match end_use:
-        case "power":  # electricity only study
-            demand_profile = demand_params["profile"]
+    if end_use == "power":  # electricity only study
+        demand_profile = demand_params["profile"]
+        demand_disaggregation = "pop"
+        if demand_profile == "efs":
+            scaling_method = "efs"
+        elif demand_profile == "eia":
+            scaling_method = "aeo_electricity"
+        elif demand_profile == "ferc":
+            scaling_method = "aeo_electricity"
+        else:
+            logger.warning(
+                f"No scaling method available for {demand_profile} profile. Setting to 'aeo_electricity'",
+            )
+    elif end_use in ("residential", "commercial"):
+        demand_profile = "eulp"
+        demand_disaggregation = "pop"
+        scaling_method = "aeo_energy"
+    elif end_use == "industry":
+        demand_profile = "cliu"
+        demand_disaggregation = "cliu"
+        scaling_method = "aeo_energy"
+    elif end_use == "transport":
+        vehicle = kwargs.get("vehicle", None)
+        if not vehicle:  # road transport
+            demand_profile = "transport_efs_aeo"
             demand_disaggregation = "pop"
-            if demand_profile == "efs":
-                scaling_method = "efs"
-            elif demand_profile == "eia":
-                scaling_method = "aeo_electricity"
-            elif demand_profile == "ferc":
-                scaling_method = "aeo_electricity"
-            else:
-                logger.warning(
-                    f"No scaling method available for {demand_profile} profile. Setting to 'aeo_electricity'",
-                )
-        case "residential" | "commercial":
-            demand_profile = "eulp"
+            scaling_method = "aeo_vmt"
+        elif vehicle.startswith(("air", "rail", "boat")):
+            demand_profile = "transport_aeo"
             demand_disaggregation = "pop"
-            scaling_method = "aeo_energy"
-        case "industry":
-            demand_profile = "cliu"
-            demand_disaggregation = "cliu"
-            scaling_method = "aeo_energy"
-        case "transport":
-            vehicle = kwargs.get("vehicle", None)
-            if not vehicle:  # road transport
-                demand_profile = "transport_efs_aeo"
-                demand_disaggregation = "pop"
-                scaling_method = "aeo_vmt"
-            elif vehicle.startswith(("air", "rail", "boat")):
-                demand_profile = "transport_aeo"
-                demand_disaggregation = "pop"
-                scaling_method = None  # will extract data for any year
-            else:
-                raise NotImplementedError
-        case _:
+            scaling_method = None  # will extract data for any year
+        else:
             raise NotImplementedError
+    else:
+        raise ValueError(end_use)
 
     return demand_profile, demand_disaggregation, scaling_method
